week02-chunking/minseon/requirements.txt
langchain
langchain-community
langchain-text-splitters
pymupdf
python-dotenv
streamlit
